+++ daemon-impl/src/main/java/ru/s4m1d/notes/app/desktop/client/daemon/context/service/ContextService.java
@@ -1,4 +1,4 @@
-package ru.s4m1d.notes.app.desktop.client.controller.context.service;
+package ru.s4m1d.notes.app.desktop.client.daemon.context.service;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
