+++ daemon-impl/src/main/java/ru/s4m1d/notes/app/desktop/client/daemon/notes/NotesBarUpdateServiceImpl.java
@@ -1,6 +1,6 @@
-package ru.s4m1d.notes.app.desktop.client.controller.notes;
+package ru.s4m1d.notes.app.desktop.client.daemon.notes;
 
-import ru.s4m1d.notes.app.desktop.client.controller.task.NewComponentEvent;
+import ru.s4m1d.notes.app.desktop.client.daemon.task.NewComponentEvent;
 import ru.s4m1d.notes.app.desktop.client.model.note.Note;
 import ru.s4m1d.notes.app.desktop.client.components.choice.bar.NoteTab;
 import ru.s4m1d.notes.app.desktop.client.components.choice.bar.NotesPane;
