+++ daemon-impl/src/main/java/ru/s4m1d/notes/app/desktop/client/daemon/notes/NotesMonitoringServiceImpl.java
@@ -1,6 +1,6 @@
-package ru.s4m1d.notes.app.desktop.client.controller.notes;
+package ru.s4m1d.notes.app.desktop.client.daemon.notes;
 
-import ru.s4m1d.notes.app.desktop.client.controller.task.NotesSearchService;
+import ru.s4m1d.notes.app.desktop.client.daemon.task.NotesSearchService;
 import ru.s4m1d.notes.app.desktop.client.model.note.Note;
 import ru.s4m1d.notes.app.desktop.client.core.Observer;
