+++ daemon-impl/src/main/java/ru/s4m1d/notes/app/desktop/client/daemon/notes/NotesSearchServiceImpl.java
@@ -1,8 +1,8 @@
-package ru.s4m1d.notes.app.desktop.client.controller.notes;
+package ru.s4m1d.notes.app.desktop.client.daemon.notes;
 
-import ru.s4m1d.notes.app.desktop.client.controller.task.NotesSearchService;
+import ru.s4m1d.notes.app.desktop.client.daemon.task.NotesSearchService;
 import ru.s4m1d.notes.app.desktop.client.model.note.Note;
-import ru.s4m1d.notes.app.desktop.client.controller.notes.mapper.FilesToNotesMapper;
+import ru.s4m1d.notes.app.desktop.client.daemon.notes.mapper.FilesToNotesMapper;
 import ru.s4m1d.notes.app.desktop.client.system.file.FileScanner;
 import ru.s4m1d.notes.app.desktop.client.system.file.FileScannerImpl;
